Remove array-init dependency, move wr/rd base instructions to segmentation submodule, fix CI
