+++ .github/workflows/build.yml
@@ -62,23 +62,26 @@ jobs:
     - name: "Run cargo build"
       run: cargo build
 
+    - name: "Run cargo build for stable without instructions"
+      run: cargo build --no-default-features
+
     - name: "Run cargo build for stable"
-      run: cargo build --no-default-features --features stable
+      run: cargo build --no-default-features --features external_asm,instructions
       if: runner.os != 'Windows'
 
     - name: "Run cargo build for stable on musl"
-      run: cargo build --target x86_64-unknown-linux-musl --no-default-features --features stable
+      run: cargo build --target x86_64-unknown-linux-musl --no-default-features --features external_asm,instructions
       if: runner.os == 'Linux'
 
     - name: "Run cargo test"
       run: cargo test
 
     - name: "Run cargo test for stable"
-      run: cargo test --no-default-features --features stable
+      run: cargo test --no-default-features --features external_asm,instructions
       if: runner.os != 'Windows'
 
     - name: "Run cargo test for stable on musl"
-      run: cargo test --target x86_64-unknown-linux-musl --no-default-features --features stable
+      run: cargo test --target x86_64-unknown-linux-musl --no-default-features --features external_asm,instructions
       if: runner.os == 'Linux'
 
     - name: 'Deny Warnings'
