+++ Cargo.toml
@@ -28,7 +28,6 @@ edition = "2018"
 [dependencies]
 bit_field = "0.9.0"
 bitflags = "1.0.4"
-array-init = { version = "0.1.1", optional = true }
 
 [build-dependencies]
 cc = { version = "1.0.37", optional = true }
@@ -37,7 +36,7 @@ cc = { version = "1.0.37", optional = true }
 default = [ "nightly", "instructions" ]
 instructions = []
 deny-warnings = []
-external_asm = [ "cc", "array-init" ]
+external_asm = [ "cc" ]
 nightly = [ "inline_asm", "const_fn", "abi_x86_interrupt" ]
 inline_asm = []
 abi_x86_interrupt = []
