+++ src/registers/mod.rs
@@ -5,7 +5,7 @@ pub mod model_specific;
 pub mod rflags;
 
 #[cfg(feature = "instructions")]
-pub use crate::instructions::{rdfsbase, rdgsbase, wrfsbase, wrgsbase};
+pub use crate::instructions::segmentation::{rdfsbase, rdgsbase, wrfsbase, wrgsbase};
 
 #[cfg(all(feature = "instructions", feature = "inline_asm"))]
 pub use crate::instructions::read_rip;
